R coding of mtcars dataset
This is my solution for the mtcars dataset which is an inbuilt dataset in RStudio. Here, the target variable is mpg(miles per gallon). After plotting all the possible variables, we have found out that wt(weight) and hp(horsepower) having the relation with output variable mpg. We are using multiple linear regression to solve this model.